+++ NEWS.md
@@ -1,4 +1,4 @@
-# tern.gee 0.0.0.9002
+# tern.gee 0.0.0.9003
 ### New features
 
 * New package